src/requirements.txt
pandas
numpy
streamlit
python-dotenv
openai